launcher: the review round — routing, blind complaints, a popup for life
The launcher slice reviewed here, and the findings fixed. Three were real
hazards; the rest were comments that had stopped being true.

- os64get.conf: `*.conf = /etc` had replaced an explicit list, and TWO of
  the names it replaced carry no .conf suffix. husk.rc and hosts were
  falling through to `* = /bin` and would have installed as programs on
  the next `os64get -a` — with /etc's copies quietly never refreshed
  again. Both spelled out, with the reason on the line so the next tidy-up
  leaves them alone. (limine.conf still escapes the glob because exact
  beats suffix; all ten files in etc/ were walked against conf_route.)

- os64_complain(): lifted out of desktop.c into libos64, and grootmenu
  uses it. The desktop learned on 2026-08-25 that a complaint on a GUI
  boot's console is a complaint nobody finds; the launcher it spawns
  inherits that same blind stderr and had TEN complaint sites and no log
  line at all. Proven with a deliberately broken /home/menu.conf: the log
  got "task 62 exited (1)" and nothing else, while the diagnosis — file,
  line and reason — sat on VT1. Now both.

- GUI_WINDOW_POPUP, latched at birth. NO_DECORATIONS plus PINNED AT CREATE
  is what makes a menu, a cascade, a tooltip, and such a window declines
  the chrome verbs: a maximized menu cannot dismiss itself, because
  dismissal is focus-loss and a window filling the screen leaves nothing
  else to click. LATCHED rather than re-read from the live flag word,
  because both halves of that pair are things a chord can change
  afterwards — reading them live got it wrong in both directions, letting
  an unpin walk straight around the guard, and letting a pinned borderless
  gterm lose its titlebar toggle. The bit crosses no boundary: client_flags
  masks it on the way in, get_state masks it on the way out.

Comments the slice had made false, each ranked as what it hid:

- window.h said the wm_ layer NEVER reads window_t.owner. focus_window
  reads and compares it — that comparison is the sibling bit the whole
  cascade depends on.
- wm_deliver_event said it runs in the COMPOSITOR's thread context, which
  was exactly true until the birth and death focus grabs began delivering
  from a client's own create/destroy syscall. That sentence is the premise
  for the invariant-4 argument, so it had to say both callers.
- wm_topmost_at said "NULL for the desktop", from before the desktop was a
  window. Its window answers for bare wallpaper now; NULL means the glass
  is truly empty.

And the smaller ones:

- focus_window: owner 0 is the not-a-task sentinel, so two windows wearing
  it are strangers, not siblings. Every other reader of `owner` already
  refuses 0 by name; this was the one place it would have meant "yes".
- The chord handler reports a verb's OUTCOME, not its intent — a line
  printed before the call left "window 2 maximized" sitting directly above
  "maximize declined".
- os64_menu_argv REFUSES rather than truncates (int64_t, negative on both
  ceilings): half a command line is a different command, and the caller
  cannot see the difference in the argv it gets back. grootmenu says which
  of the two happened, and says when a level ran past MENU_ROWS_MAX.
- The menu parser names a stray `}` where a cascade's reference belongs,
  instead of eating it and refusing the file later for the wrong reason at
  the wrong line.
- shutdown: "no tasks to stop" and the tick-clock-stall notice reach the
  log like the other verdicts. The DEBUG_BOOT note names its reason (the
  Ctrl+~ suppression leaves that bit on) instead of claiming "always".
- theme.conf's menu.border had no consumer — the panel is menu.bg and the
  1px edge is the WM's — so the key, its default and its doc line are
  gone rather than silently doing nothing.

Verified on a headless GUI boot, driven by hand: menu at the click,
inline cascade and cascade-by-reference with the parent standing, Terminal
and Tools->Top both launching, dismissal by click-away, by a stranger
window and by Escape, exit 0 to the log alone. Both popup holes retested
and closed — an unpinned menu still declines all three verbs, and a
stripped-and-pinned gterm keeps all three. Clean descent, no faults.

+++ userland/libos64/include/os64/menu.h
@@ -119,9 +119,12 @@ bool os64_menu_named_exists(const os64_menu_t *menu, const char *name);
 
 // Split an item's command line into an argv for os64_spawn: whitespace
 // separates, double quotes group, `buf` receives the NUL-separated words
-// and `argv` the pointers (argv[n] = NULL). Returns the word count, or 0
-// for an empty command or one that would not fit.
-size_t os64_menu_argv(const char *command, char *buf, size_t buf_cap,
-                      char *argv[], size_t argv_max);
+// and `argv` the pointers (argv[n] = NULL). Returns the word count, 0 for
+// an empty command, or negative when the words do not fit in `buf` or
+// `argv` — a REFUSAL, not a truncation, because half a command line is a
+// different command and the caller cannot see the difference in the argv it
+// gets back.
+int64_t os64_menu_argv(const char *command, char *buf, size_t buf_cap,
+                       char *argv[], size_t argv_max);
 
 #endif // OS64_MENU_H
